% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{ngram}
\alias{ngram}
\title{Get Google Books Ngram Data}
\usage{
ngram(terms, yr.start = 1800, yr.end = 2000, smoothing = 0, corpus = 15)
}
\arguments{
\item{terms}{Terms to query against the Google Books corpus. Typically a vector.}

\item{yr.start}{The first year to query. Defaults to 1800.}

\item{yr.end}{The last year to query. Defaults to 2008.}

\item{smoothing}{The number of years to smooth over. Defaults to 0.}

\item{corpus}{The corpus to query. Defaults to 15, which corresponds to the English language corpus.}
}
\description{
This function fetches the underlying plot data for Google Books Ngrams. It returns a dataframe with the frequency of each term for each year within a specified range.
}
\examples{

library(ggplot2)
library(tidyr)
library(dplyr)

# set query terms
main.terms <- c("soldier", "politician", "writer")

# plot by term
df <- ngram(main.terms)

df.plot <- df \%>\%
  gather(term, frequency, -year)

p <- ggplot(df.plot, aes(year, frequency, colour=term)) + geom_line() +
  ylab("Frequency") +
  xlab("Year") +
  ggtitle("Terms in Google Books, 1800-2000") +
  theme_bw()
p
}
\keyword{ngrams}

